+++ app/exceptions/item_not_found_error.rb
@@ -1,6 +1,6 @@
 class ItemNotFoundError < StandardError
   attr_reader :message
-
+  
   def initialize(message)
   @message = message
   end
